chore: rename useSearchParam to useQueryParam
